+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ## [Unreleased]
 ### Added
+- add .aiignore to template
 - gemfiles/modular/x_std_libs pattern to template, including:
   - erb
   - mutex_m